server/requirements.txt
datasets 
# vmdpy
scikit-learn==1.6.1
duckdb
matplotlib
flask
numpy
flask_cors
python-dotenv
gunicorn
joblib
tensorflow
torch